requirements.txt
pydantic-ai
open-interpreter
docker
sqlalchemy
aiosqlite
pydantic-settings
pytest
chainlit
python-dotenv
